backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
openai
python-dotenv
pydantic
psycopg2-binary
requests
